+++ application/controllers/Player_Portfolio.php
@@ -1,0 +1,14 @@
+<?php
+
+class Player_Portfolio extends MY_Controller {
+    function __construct()
+    {
+        parent::__construct();
+    }
+
+    function index() {
+        $this->data['pagetitle'] = 'Player Portfolio';
+        $this->data['pagebody'] = 'player_portfolio';
+        $this->render();
+    }
+}